Directory support, Refactor of Json -> .jsonparser | Starting wrapper classes
Json difficult to work with, so am just going to create the actual object classes separate from the json data so I can be lazy and use as much gson default implementation as possible. Looked at Type Adaptors for a bit before deciding against them. I'll probably figure out that I could have just used maps or something instead of hardcoding the objects...but thats something to come to later. Oh and I deleted the .idea stuff that wasn't meant to be in the repo.

Guess this is the result of init repo in IntelliJ